Size -> Vector2f

+++ src/main/java/mcp/mobius/waila/api/ui/IDisplayHelper.java
@@ -6,11 +6,11 @@ import net.minecraft.item.ItemStack;
 
 public interface IDisplayHelper {
 
-	void drawItem(MatrixStack matrixStack, int x, int y, ItemStack stack, float scale);
+	void drawItem(MatrixStack matrixStack, float x, float y, ItemStack stack, float scale);
 
 	void drawGradientRect(MatrixStack matrixStack, float left, float top, float right, float bottom, int startColor, int endColor);
 
-	void drawBorder(MatrixStack matrixStack, int minX, int minY, int maxX, int maxY, IBorderStyle border);
+	void drawBorder(MatrixStack matrixStack, float minX, float minY, float maxX, float maxY, IBorderStyle border);
 
 	String humanReadableNumber(double number, String unit, boolean milli);
